Update README.md
About Number Guessing Game Project

This game is about creating an interactive game that will display scores and a person has to guess the number. There will be hints displayed every time the user guesses a wrong number. The objective of creating this project is to challenge our brains and add some fun while learning python.